fix(icon): restore 1024x1024 Vangard icon and rename from vangard-512x512.png to vangard.png
The correct 1024x1024 icon was never committed — it existed only as a local working-copy change that was accidentally dropped during a stash operation. The old committed file was a 512x512 placeholder causing the Electron default icon to appear. Renamed to vangard.png to reflect the actual dimensions and update all references (electron.js, package.json, Toolbar.tsx, AboutModal.tsx, Toolbar.test.tsx).
